Move dashboard components to dashboard directory
